Got Stat Endpoints into separate folder, and moved classes over
Plus some cleanup
