--- api/ExpressedRealms.Characters.API/CharacterEndPoints/StatDTOs/EditStatRequest.cs
@@ -1,5 +1,0 @@
-using ExpressedRealms.Characters.Repository.Stats.Enums;
-
-namespace ExpressedRealms.Characters.API.CharacterEndPoints.StatDTOs;
-
-internal record EditStatRequest(int CharacterId, StatType StatTypeId);
