+++ api/ExpressedRealms.Characters.API/Configuration/CharactersAPIConfiguration.cs
@@ -1,11 +1,12 @@
 using ExpressedRealms.Characters.API.CharacterEndPoints;
+using ExpressedRealms.Characters.API.StatEndPoints;
 using Microsoft.AspNetCore.Builder;
 
 namespace ExpressedRealms.Characters.API.Configuration;
 
-public static class CharactersAPIConfiguration
+public static class CharactersApiConfiguration
 {
-    public static void AddCharacterAPIEndPoints(this WebApplication app)
+    public static void AddCharacterApiEndPoints(this WebApplication app)
     {
         app.AddCharacterEndPoints();
         app.AddStatEndPoints();
