+++ api/ExpressedRealms.Characters.API/StatEndPoints/Responses/SingleStatInfo.cs
