+++ api/ExpressedRealms.Characters.API/StatEndPoints/Responses/SmallStatInfo.cs
