+++ api/ExpressedRealms.Characters.API/StatEndPoints/Responses/StatDetails.cs
